+++ PingIt/Core/Protocols/FirebaseUser+AuthUserRepresentable.swift
@@ -1,0 +1,3 @@
+import FirebaseAuth
+
+extension FirebaseAuth.User: AuthUserRepresentable {}