+++ PingItTests/Mocks/MockAuthUser.swift
@@ -1,0 +1,5 @@
+@testable import PingIt
+
+struct MockAuthUser: AuthUserRepresentable {
+    var uid: String
+}